Add the licence type to the top of the example license
Whilst the `licence_info.md` content explained that the license is
the BSD 3-Clause licence, the actual license template copy did not.

Without the name of the licence type it may be harder for contributors
to an open source project to know if the licence is compatible with
any policies their employer may have, unless they are able to recognise
the licence by its content.

In addition, I've changed the bullets to numbers, given:
(a) all examples of the BSD 3-Clause licence on other sites use numbers
(b) it drives home the "3 clause" aspect.