Make CONTRIBUTING.md copy/paste-able
The default CONTRIBUTING.md requires quite a few modifications before it can be used. Instead of requiring that the developer do this correctly the first time, this file is designed to be correct and ready to use as-is, but also contains suggested content that can be gradually rolled out over time by removing HTML comments.

Here's a rendered version of this page in GitHub markdown https://gist.github.com/schneems/042d1925ad5d6cadc278b8f028c9aef6

Note the things like ` <!-- For {NAME OF PROJECT} -->` are not rendered as they're considered an html comment.